Support for specifying direction of strings. Fixes #1
play and tune now accept :low_to_high or :high_to_low as a second optional argument.
This argument specifies direction, and is meaningful only when used with arrays.
The default is :high_to_low.

A Chord instance can now be reversed with the reverse_strings! function.

Improved exception handling in chordy.rb.